+++ apps/onyx/lib/core/extensions/extensions_export.dart
@@ -1,10 +1,10 @@
-export 'duration_beautifull.dart';
-export 'date_shrink.dart';
-export 'functionalities_to_human_export.dart';
-export 'hex.dart';
-export 'int_to_sized_string.dart';
-export 'lat_lng_inside.dart';
-export 'month_to_string.dart';
-export 'presence_color_to_flutter_color.dart';
-export 'weekday_to_string.dart';
-export 'date_beautifull.dart';
+export 'duration_extension.dart';
+export 'date_extension.dart';
+export 'functionalities_extension.dart';
+export 'color_extension.dart';
+export 'coords_extension.dart';
+export 'int_extension.dart';
+export 'lat_lng_extension.dart';
+export 'presence_color_extension.dart';
+export 'special_mail_box_extension.dart';
+export 'mail_box_extension.dart';
