+++ apps/onyx/lib/core/extensions/mail_box_extension.dart
@@ -1,23 +1,5 @@
-import 'package:enough_mail/enough_mail.dart';
 import 'package:flutter/material.dart';
 import 'package:lyon1mailclient/lyon1mailclient.dart';
-
-extension MailBoxTag on SpecialMailBox {
-  MailboxFlag toMailBoxTag() {
-    switch (this) {
-      case SpecialMailBox.inbox:
-        return MailboxFlag.inbox;
-      case SpecialMailBox.sent:
-        return MailboxFlag.sent;
-      case SpecialMailBox.trash:
-        return MailboxFlag.trash;
-      case SpecialMailBox.flagged:
-        return MailboxFlag.flagged;
-      case SpecialMailBox.archive:
-        return MailboxFlag.archive;
-    }
-  }
-}
 
 extension MailBoxExtension on MailBox {
   IconData toIcon() {
