+++ apps/onyx/lib/core/extensions/presence_color_extension.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:lyon1tomussclient/lyon1tomussclient.dart';
 import 'package:onyx/core/theme/theme_export.dart';
 
-extension PresenceColorConverter on PresenceColor {
+extension PresenceColorExtension on PresenceColor {
   Color toColor() {
     switch (this) {
       case PresenceColor.green:
