+++ apps/onyx/lib/screens/agenda/agenda_export.dart
@@ -10,3 +10,4 @@ export 'widgets/one_day_view_widget.dart';
 export 'widgets/multiple_day_view/multiple_day_view_widget.dart';
 export 'widgets/multiple_day_view/left_hour_indicator_widget.dart';
 export 'widgets/multiple_day_view/grid_widget.dart';
+export 'widgets/multiple_day_view/top_day_indicator.dart';
