+++ apps/onyx/lib/screens/mails/logic/email_logic.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:lyon1mailclient/lyon1mailclient.dart';
 import 'package:onyx/core/cache_service.dart';
-import 'package:onyx/core/extensions/mail_box_extension.dart';
+import 'package:onyx/core/extensions/special_mail_box_extension.dart';
 import 'package:onyx/core/initialisations/initialisations_export.dart';
 import 'package:onyx/core/res.dart';
